requirements.txt
django==1.8.6
djangorestframework==3.3.2
requests==2.9.1
beautifulsoup4
mysql-python
PIL

<canshu cuowuchuli>
<space, zhushi>
<soup not found>
<list import better>
<meituan shoujitehu>
<join>

<token>
<like meituan currentpage jiancha>

